fix region

+++ src/modules/regionarea/service/region.service.ts
@@ -23,12 +23,15 @@ export class RegionService {
     try {
       createDto.name = createDto.name.toUpperCase();
       const findRegion = await this.RegionRepository.getByName(createDto.name);
-      if (findRegion[0].deleted_at !== null) {
-        await this.RegionRepository.deleteForce(findRegion[0].id);
-      }else{
-        throw new HttpException('Region already exists', 400);
+      if (findRegion[0]) {
+        if (findRegion[0].deleted_at !== null) {
+          await this.RegionRepository.deleteForce(findRegion[0].id);
+        } else {
+          throw new HttpException('Region already exists', 400);
+        }
+      } else {
+        return await this.RegionRepository.create(createDto);
       }
-      return await this.RegionRepository.create(createDto);
     } catch (e) {
       logger.error('Error in create data:', e.message, e.stack);
       if (e instanceof HttpException) {
